LargeWord.toIntX deal with conversion to signed int incorrectly. Fixes aliceml/aliceml#6.
This is caused by type conversions in the implementation of fromIntX in C++.
It converts the value passed to toIntX to an uint32_t. This is compared
to MIN/MAX int values by converting it to a signed long int. It then compares
as greater than the MAX int value. Fix is to cast to a int32_t before
comparing as a signed long int.

+++ vm-seam/alice/primitives/Word32.cc
@@ -126,10 +126,10 @@ DEFINE1(Word32_toInt) {
 
 DEFINE1(Word32_toIntX) {
   DECLARE_WORD32(i, x0);
-  if (static_cast<s_int>(i) > static_cast<s_int>(MAX_VALID_INT) ||
-      static_cast<s_int>(i) < static_cast<s_int>(MIN_VALID_INT))
+  if (static_cast<s_int>(static_cast<int32_t>(i)) > static_cast<s_int>(MAX_VALID_INT) ||
+      static_cast<s_int>(static_cast<int32_t>(i)) < static_cast<s_int>(MIN_VALID_INT))
     RAISE(PrimitiveTable::General_Overflow);
-  RETURN_INT(i);
+  RETURN_INT(static_cast<int32_t>(i));
 } END
 
 DEFINE1(Word32_fromLargeInt) {
